requirements.txt
appdirs==1.4.3
packaging==16.8
pygame==1.9.3
PyOpenGL==3.1.0
pyparsing==2.2.0
six==1.10.0
